Separate deltadb_query_main from primary query object. Still needs some cleanup
