+++ deltadb/src/Makefile
@@ -6,14 +6,16 @@ LIBRARIES = libdeltadb.a
 OBJECTS = $(SOURCES:%.c=%.o)
 PROGRAMS = deltadb_query deltadb_upgrade_log
 SCRIPTS =
-SOURCES = deltadb_stream.c deltadb_reduction.c
+SOURCES = deltadb_query.c deltadb_stream.c deltadb_reduction.c
 TARGETS = $(LIBRARIES) $(PROGRAMS)
 
 all: $(TARGETS)
 
 libdeltadb.a: $(OBJECTS)
 
-$(PROGRAMS): $(LIBRARIES) $(EXTERNAL_DEPENDENCIES)
+deltadb_query: deltadb_query_main.o libdeltadb.a $(EXTERNAL_DEPENDENCIES)
+
+deltadb_upgrade_log: deltadb_upgrade_log.o libdeltadb.a $(EXTERNAL_DEPENDENCIES)
 
 clean:
 	rm -f $(OBJECTS) $(TARGETS) *.o
